Created a Combat folder under Rules. Added relavent existing classes. After seeing smartypan's folder restructure, I see a way forward in organizing. The intent is to capture relevant rules in the same folder tree/branch.
